+++ capycli/bom/handle_bom.py
@@ -15,6 +15,7 @@ import capycli.bom.check_granularity
 import capycli.bom.create_components
 import capycli.bom.diff_bom
 import capycli.bom.download_sources
+import capycli.bom.download_attachments
 import capycli.bom.filter_bom
 import capycli.bom.findsources
 import capycli.bom.map_bom
@@ -35,19 +36,20 @@ def run_bom_command(args) -> None:
 
         # display `bom` related help
         print("bom               bill of material (BOM) specific sub-commands")
-        print("    Show              display contents of a BOM")
-        print("    Convert           Convert SBOM formats")
-        print("    Filter            apply filter file to a BOM")
-        print("    Check             check that all releases in the BOM exist on target SW360 instance")
-        print("    CheckItemStatus   show additional information about BOM items on SW360")
-        print("    Map               map a given BOM to data on SW360")
-        print("    CreateReleases    create new releases for existing components on SW360")
-        print("    CreateComponents  create new components and releases on SW360 (use with care!)")
-        print("    DownloadSources   download source files from the URL specified in the SBOM")
-        print("    Granularity       check a bill of material for potential component granularity issues")
-        print("    Diff              compare two bills of material.")
-        print("    Merge             merge two bills of material.")
-        print("    Findsources       determine the source code for SBOM items.")
+        print("    Show                display contents of a BOM")
+        print("    Convert             Convert SBOM formats")
+        print("    Filter              apply filter file to a BOM")
+        print("    Check               check that all releases in the BOM exist on target SW360 instance")
+        print("    CheckItemStatus     show additional information about BOM items on SW360")
+        print("    Map                 map a given BOM to data on SW360")
+        print("    CreateReleases      create new releases for existing components on SW360")
+        print("    CreateComponents    create new components and releases on SW360 (use with care!)")
+        print("    DownloadAttachments download SW360 attachments as specified in the SBOM")
+        print("    DownloadSources     download source files from the URL specified in the SBOM")
+        print("    Granularity         check a bill of material for potential component granularity issues")
+        print("    Diff                compare two bills of material.")
+        print("    Merge               merge two bills of material.")
+        print("    Findsources         determine the source code for SBOM items.")
         return
 
     subcommand = args.command[1].lower()
@@ -97,6 +99,12 @@ def run_bom_command(args) -> None:
     if subcommand == "downloadsources":
         """Download source files from the URL specified in the SBOM."""
         app = capycli.bom.download_sources.BomDownloadSources()
+        app.run(args)
+        return
+
+    if subcommand == "downloadattachments":
+        """Download attachments from SW360 as specified in the SBOM."""
+        app = capycli.bom.download_attachments.BomDownloadAttachments()
         app.run(args)
         return